Modify template file
Add a template file that's a small modification of the upstream default.
The modification is required because Pulp uses hrefs (which include
multiple path components) as path parameters to API calls and they
should not be escaped.

The following is the diff between our template and the current default:

127c127
< 	localVarPath = strings.Replace(localVarPath, "{"+"{{baseName}}"+"}", url.PathEscape(parameterValueToString(r.{{paramName}}, "{{paramName}}")), -1){{/pathParams}}
---
> 	localVarPath = strings.Replace(localVarPath, "{"+"{{baseName}}"+"}", parameterValueToString(r.{{paramName}}, "{{paramName}}"), -1)  // NOTE: paths aren't escaped because Pulp uses hrefs as path parameters{{/pathParams}}

(The url.PathEscape() call is removed)